Refactor transaction history to use `TransactionResponse`
Updated `TransactionHistoryService` and related tests to return `TransactionResponse` instead of `Transaction`. This ensures a clear separation between internal models and API response objects, streamlining data handling and improving maintainability.
