+++ src/test/java/info/mackiewicz/bankapp/integration/TransactionHistoryRestControllerIntegrationTest.java
@@ -1,10 +1,10 @@
 package info.mackiewicz.bankapp.integration;
 
 import info.mackiewicz.bankapp.account.model.Account;
-import info.mackiewicz.bankapp.system.banking.history.dto.TransactionFilterRequest;
+import info.mackiewicz.bankapp.system.banking.history.controller.dto.TransactionFilterRequest;
 import info.mackiewicz.bankapp.system.banking.history.service.TransactionHistoryService;
+import info.mackiewicz.bankapp.system.banking.shared.dto.TransactionResponse;
 import info.mackiewicz.bankapp.testutils.TestAccountBuilder;
-import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.user.model.User;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
@@ -58,7 +58,7 @@ class TransactionHistoryRestControllerIntegrationTest {
 
     @BeforeEach
     void setUp() {
-        Page<Transaction> transactionPage = new PageImpl<>(Collections.emptyList());
+        Page<TransactionResponse> transactionPage = new PageImpl<>(Collections.emptyList());
 
         when(transactionHistoryService.getTransactionHistory(any(TransactionFilterRequest.class)))
                 .thenReturn(transactionPage);
